requirements.txt
flask==3.0.3
flask_cors==5.0.0
bbc-news==1.0.2